chore(assets): Move Assets outside of UI directory
IMO Assets does not belong to the UI directory as UI directory should
contain only components. Moreover Assets can be consumed not only by UI
components.
