+++ src/assets/svg/index.ts
