+++ src/ui/components/app-shortcuts/template.ts
@@ -1,7 +1,7 @@
 import { html, TemplateResult } from 'lit-element'
 import { AppShortcutsElement } from './component'
 import { repeat } from 'lit-html/directives/repeat'
-import { Icon } from '../../assets/svg'
+import { Icon } from '../../../assets/svg'
 
 export default function template(this: AppShortcutsElement): TemplateResult {
   return html`
