+++ src/ui/containers/main-launcher/component.ts
@@ -17,7 +17,7 @@ import {
   startApplication,
   selectApplication,
 } from '../../../store'
-import { Theme } from '../../assets/style'
+import { Theme } from '../../../assets/style'
 import template from './template'
 import style from './style'
 import { NavItemElementData } from '../../components/nav-item'
