+++ src/ui/views/app-home/component.ts
@@ -1,6 +1,6 @@
 import { LitElement, property, customElement, TemplateResult, CSSResultArray } from 'lit-element'
 import { store, connect, RootState, WebsocketSelectors, Application } from '../../../store'
-import { Theme } from '../../assets/style'
+import { Theme } from '../../../assets/style'
 import template from './template'
 import style from './style'
